+++ packages/charts/src/chart_types/xy_chart/utils/specs.ts
@@ -878,11 +878,6 @@ export type LineAnnotationSpec<D = any> = BaseAnnotationSpec<
   /** Annotation lines are hidden */
   hideLines?: boolean;
   /**
-   * Hide tooltip when hovering over the line
-   * @defaultValue `true`
-   */
-  hideLinesTooltips?: boolean;
-  /**
    * z-index of the annotation relative to other elements in the chart
    * @defaultValue 1
    */